minor refactor, separate embed creation to a separated file, also learned how to destructure dict into a list of params
